requirements.txt
customtkinter
requests
mysql-connector-python==8.0.26
pandas
Pillow
